NX-34: close external-reference string-use slot positions
Replace the unrestricted native slot byte with the four ExtrefSlot variants.
Extraction visits those variants directly and removes the usize-to-u8 cast.
Slot byte offsets derive from the handle-set layout ID_SLOTS constant.
Numeric JSON is unchanged; deserialization rejects values outside 0..=3.
Existing complete-lane and source-offset assertions are unchanged.

The child join still checks that four independently supplied occurrences
cover all four positions exactly once. That is a cross-record completeness
rule, not a possible invalid value of the new slot type. The wider NX-34
audit remains open. cargo check -q -p cadmpeg-codec-nx --tests passes.

Hook-Bypass-Reason: The user requires scoped compile checks between items and defers the full gate to the completed refactor.

+++ crates/cadmpeg-codec-nx/src/native/om/tests/material_and_external_records.rs
@@ -530,7 +530,7 @@ fn external_reference_record_slots_resolve_atomically_in_the_same_stream() {
     assert_eq!(uses.len(), 4);
     assert_eq!(uses[0].id, "nx:external-reference:record-string-use#7-0");
     assert_eq!(
-        uses.iter().map(|use_| use_.slot).collect::<Vec<_>>(),
+        uses.iter().map(|use_| u8::from(use_.slot)).collect::<Vec<_>>(),
         [0, 1, 2, 3]
     );
     assert_eq!(
